requirements.txt
# FastAPI & Web Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Computer Vision & Face Detection
insightface>=0.7.3
onnxruntime>=1.20.0  # Updated for Python 3.13+
opencv-python>=4.10.0
opencv-contrib-python>=4.10.0
Pillow>=11.0.0

# DeepFace for Emotion Detection
deepface==0.0.95
tf-keras==2.20.1
tensorflow==2.20.0

# Machine Learning & Clustering
numpy>=1.26.0,<3.0.0
scikit-learn>=1.5.0
scipy>=1.14.0
hdbscan>=0.8.38

# Image Processing Utilities
imageio>=2.36.0
colorthief>=0.2.1

# File & Path Management
aiofiles>=24.1.0
python-magic>=0.4.27

# Logging & Monitoring
loguru>=0.7.3

# AI Integration
openai>=1.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
google-generativeai>=0.8.0
